requirements.txt
# Main tendenci repo
git+https://github.com/tendenci/tendenci#egg=tendenci

# Heroku deployment packages
dj-database-url==0.2.1
psycopg2==2.4.5
gunicorn==0.14.5
django-s3-folder-storage==0.1
django-storages==1.1.5

# Add in Geraldo
git+https://github.com/tendenci/geraldo#egg=geraldo


# The following packages are needed for Heroku.
# You may not need them locally.

# Memcache
pylibmc==1.2.2
django-pylibmc-sasl==0.2.4

# For Sentry
raven==2.0.2
